+++ packages/client/src/tx/create-tx.ts
@@ -1,15 +1,11 @@
-import { Observable, combineLatest, mergeMap, of, take } from "rxjs"
+import { HexString, u16, u32, u64, u8 } from "@polkadot-api/substrate-bindings"
+import { Observable, combineLatest, map, mergeMap, of, take } from "rxjs"
 import { BlockInfo, ChainHead$ } from "@polkadot-api/observable-client"
-import {
-  ChargeAssetTxPayment,
-  ChargeTransactionPayment,
-  CheckMortality,
-} from "./signed-extensions/user"
-import * as chainSignedExtensions from "./signed-extensions/chain"
 import type { PolkadotSigner } from "@polkadot-api/polkadot-signer"
 import { _void } from "@polkadot-api/substrate-bindings"
 import { CustomSignedExtensionValues } from "./types"
-import { mapObject } from "@polkadot-api/utils"
+import { fromHex, toHex } from "@polkadot-api/utils"
+import { getSignExtensionsCreator } from "./signed-extensions"
 
 type HintedSignedExtensions = Partial<{
   tip: bigint
@@ -17,6 +13,29 @@ type HintedSignedExtensions = Partial<{
   asset: Uint8Array
   nonce: number
 }>
+
+const NONCE_RUNTIME_CALL = "AccountNonceApi_account_nonce"
+const lenToDecoder = {
+  1: u8.dec,
+  2: u16.dec,
+  4: u32.dec,
+  8: u64.dec,
+}
+
+export const getNonce$ = (
+  call$: ChainHead$["call$"],
+  from: HexString,
+  at: string,
+) =>
+  call$(at, NONCE_RUNTIME_CALL, from).pipe(
+    map((result) => {
+      const bytes = fromHex(result)
+      const decoder = lenToDecoder[bytes.length as 2 | 4 | 8]
+      if (!decoder)
+        throw new Error(`${NONCE_RUNTIME_CALL} retrieved wrong data`)
+      return decoder(bytes)
+    }),
+  )
 
 export const createTx: (
   chainHead: ChainHead$,
@@ -33,89 +52,45 @@ export const createTx: (
   customSignedExtensions,
   hinted = {},
 ) =>
-  chainHead.getRuntimeContext$(atBlock.hash).pipe(
+  combineLatest([
+    hinted.nonce
+      ? of(hinted.nonce)
+      : getNonce$(chainHead.call$, toHex(signer.publicKey), atBlock.hash),
+    chainHead.getRuntimeContext$(atBlock.hash),
+    chainHead.genesis$,
+  ]).pipe(
     take(1),
-    mergeMap((ctx) => {
-      const signedExtensionsCtx = {
-        lookupFn: ctx.lookup,
-        dynamicBuilder: ctx.dynamicBuilder,
-        chainHead: chainHead,
-        callData: callData,
-        at: atBlock.hash,
-        from: signer.publicKey,
-      }
+    mergeMap(([nonce, ctx, genesis]) => {
+      const signExtCreator = getSignExtensionsCreator(
+        fromHex(genesis),
+        ctx.lookup,
+        ctx.dynamicBuilder,
+      )
 
-      const mortality: Parameters<typeof CheckMortality>[0] = !hinted.mortality
-        ? { period: 64, blockNumber: atBlock.number }
-        : hinted.mortality.mortal
-          ? { period: hinted.mortality.period, blockNumber: atBlock.number }
-          : undefined // immortal
+      const mortality: HintedSignedExtensions["mortality"] =
+        hinted.mortality ?? { period: 64, mortal: true }
 
-      return combineLatest(
-        Object.fromEntries(
-          ctx.lookup.metadata.extrinsic.signedExtensions
-            .map(({ identifier, type, additionalSigned }) => {
-              const stream = () => {
-                if (identifier === "CheckMortality")
-                  return CheckMortality(mortality, signedExtensionsCtx)
+      const signExtensions = signExtCreator({
+        nonce: nonce as number,
+        tip: hinted.tip ?? 0n,
+        mortality: mortality.mortal
+          ? {
+              mortal: true,
+              period: mortality.period,
+              startAtBlock: {
+                height: atBlock.number,
+                hash: atBlock.hash,
+              },
+            }
+          : { mortal: false },
+        customSignedExtensions,
+      })
 
-                if (identifier === "ChargeTransactionPayment")
-                  return ChargeTransactionPayment(hinted.tip ?? 0n)
-
-                if (identifier === "ChargeAssetTxPayment")
-                  return ChargeAssetTxPayment(hinted.tip ?? 0n, hinted.asset)
-
-                if (identifier === "CheckNonce" && "nonce" in hinted)
-                  return chainSignedExtensions.getNonce(hinted.nonce!)
-
-                const fn = chainSignedExtensions[identifier as "CheckGenesis"]
-                if (fn) return fn(signedExtensionsCtx)
-
-                // If we reached this point, that means that it's either an "empty" or a "custom" signed-extension
-                const customEntry = customSignedExtensions[identifier] as any
-                const [[valueEnc], [additionalSignedEnc]] = [
-                  type,
-                  additionalSigned,
-                ].map(ctx.dynamicBuilder.buildDefinition)
-                try {
-                  return of(
-                    mapObject(
-                      {
-                        value: valueEnc,
-                        additionalSigned: additionalSignedEnc,
-                      },
-                      (encoder, key) => {
-                        const input = customEntry?.[key]
-                        // if the encoder is _void, then the input value is ignored, so no harm in passing `undefined`
-                        // only an `Option` encoder will accept `undefined` as an input without crashing
-                        return input instanceof Uint8Array
-                          ? input
-                          : encoder(input)
-                      },
-                    ),
-                  )
-                } catch {
-                  // this means that a non optional custom signed-extension has not received its value
-                  return null
-                }
-              }
-
-              return [identifier, stream()!] as const
-            })
-            .filter((x) => x[1]),
-        ),
-      ).pipe(
-        mergeMap((signedExtensions) =>
-          signer.signTx(
-            callData,
-            mapObject(signedExtensions, (v, identifier: string) => ({
-              identifier,
-              ...v,
-            })),
-            ctx.metadataRaw,
-            atBlock.number,
-          ),
-        ),
+      return signer.signTx(
+        callData,
+        signExtensions,
+        ctx.metadataRaw,
+        atBlock.number,
       )
     }),
   )