+++ packages/client/src/tx/signed-extensions/chain/CheckSpecVersion.ts
@@ -1,6 +1,0 @@
-import { map } from "rxjs"
-import type { GetChainSignedExtension } from "../internal-types"
-import { additionalSigned, systemVersionProp$ } from "../utils"
-
-export const CheckSpecVersion: GetChainSignedExtension = ({ lookupFn }) =>
-  systemVersionProp$("spec_version", lookupFn).pipe(map(additionalSigned))