+++ packages/client/src/tx/signed-extensions/charge-asset-tx-enc.ts
@@ -1,0 +1,11 @@
+import {
+  Bytes,
+  compact,
+  Option,
+  Struct,
+} from "@polkadot-api/substrate-bindings"
+
+export const [ChargeAssetTxPaymentEnc] = Struct({
+  tip: compact,
+  asset: Option(Bytes(Infinity)),
+})