+++ packages/client/CHANGELOG.md
@@ -7,6 +7,8 @@
 - `getOfflineApi` which mainly facilitates the creation of extrinsics while being offline.
 - **Codegen:**
   - Add XCM v5 known types
+- **Chains:**
+  - Add Polkadot coretime chainspec
 
 ### Fixed